requirements.txt
Flask==2.2.5
pandas==2.0.3
prometheus_client==0.17.1
psutil==5.9.5
python-dateutil==2.8.2
PyYAML==6.0
Requests==2.31.0

python-dotenv~=1.0.1
